feat: agregar hook useDiagnostico para gestionar síntomas en la vista de diagnóstico
